// shared/contracts/order.d.ts
export declare const OrderPayload: z.ZodObject<{
    type: z.ZodOptional<z.ZodEnum<["online", "kiosk", "voice", "drive-thru", "dine-in", "dine_in", "takeout", "delivery", "pickup"]>>;
    items: z.ZodArray<z.ZodObject<{
        id: z.ZodString;
        menu_item_id: z.ZodString;
        name: z.ZodString;
        quantity: z.ZodNumber;
        price: z.ZodNumber;
        subtotal: z.ZodOptional<z.ZodNumber>;
        modifiers: z.ZodOptional<z.ZodArray<z.ZodAny, "many">>;
        modifications: z.ZodOptional<z.ZodArray<z.ZodAny, "many">>;
        notes: z.ZodOptional<z.ZodString>;
        special_instructions: z.ZodOptional<z.ZodString>;
        specialInstructions: z.ZodOptional<z.ZodString>;
    }, "strip", z.ZodTypeAny, {
        id: string;
        name: string;
        quantity: number;
        menu_item_id: string;
        price: number;
        modifications?: any[] | undefined;
        specialInstructions?: string | undefined;
        notes?: string | undefined;
        subtotal?: number | undefined;
        modifiers?: any[] | undefined;
        special_instructions?: string | undefined;
    }, {
        id: string;
        name: string;
        quantity: number;
        menu_item_id: string;
        price: number;
        modifications?: any[] | undefined;
        specialInstructions?: string | undefined;
        notes?: string | undefined;
        subtotal?: number | undefined;
        modifiers?: any[] | undefined;
        special_instructions?: string | undefined;
    }>, "many">;
    customer_name: z.ZodOptional<z.ZodString>;
    customerName: z.ZodOptional<z.ZodString>;
    customer_email: z.ZodOptional<z.ZodString>;
    customerEmail: z.ZodOptional<z.ZodString>;
    customer_phone: z.ZodOptional<z.ZodString>;
    customerPhone: z.ZodOptional<z.ZodString>;
    table_number: z.ZodOptional<z.ZodUnion<[z.ZodString, z.ZodNumber]>>;
    tableNumber: z.ZodOptional<z.ZodUnion<[z.ZodString, z.ZodNumber]>>;
    seat_number: z.ZodOptional<z.ZodNumber>;
    seatNumber: z.ZodOptional<z.ZodNumber>;
    notes: z.ZodOptional<z.ZodString>;
    subtotal: z.ZodOptional<z.ZodNumber>;
    tax: z.ZodOptional<z.ZodNumber>;
    tip: z.ZodOptional<z.ZodNumber>;
    total_amount: z.ZodOptional<z.ZodNumber>;
    totalAmount: z.ZodOptional<z.ZodNumber>;
    metadata: z.ZodOptional<z.ZodRecord<z.ZodString, z.ZodAny>>;
}, "strip", z.ZodTypeAny, {
    items: {
        id: string;
        name: string;
        quantity: number;
        menu_item_id: string;
        price: number;
        modifications?: any[] | undefined;
        specialInstructions?: string | undefined;
        notes?: string | undefined;
        subtotal?: number | undefined;
        modifiers?: any[] | undefined;
        special_instructions?: string | undefined;
    }[];
    type?: "voice" | "online" | "pickup" | "delivery" | "dine-in" | "takeout" | "drive-thru" | "kiosk" | "dine_in" | undefined;
    metadata?: Record<string, any> | undefined;
    notes?: string | undefined;
    subtotal?: number | undefined;
    tableNumber?: string | number | undefined;
    tax?: number | undefined;
    tip?: number | undefined;
    totalAmount?: number | undefined;
    table_number?: string | number | undefined;
    customer_name?: string | undefined;
    customer_phone?: string | undefined;
    customer_email?: string | undefined;
    seat_number?: number | undefined;
    total_amount?: number | undefined;
    seatNumber?: number | undefined;
    customerName?: string | undefined;
    customerEmail?: string | undefined;
    customerPhone?: string | undefined;
}, {
    items: {
        id: string;
        name: string;
        quantity: number;
        menu_item_id: string;
        price: number;
        modifications?: any[] | undefined;
        specialInstructions?: string | undefined;
        notes?: string | undefined;
        subtotal?: number | undefined;
        modifiers?: any[] | undefined;
        special_instructions?: string | undefined;
    }[];
    type?: "voice" | "online" | "pickup" | "delivery" | "dine-in" | "takeout" | "drive-thru" | "kiosk" | "dine_in" | undefined;
    metadata?: Record<string, any> | undefined;
    notes?: string | undefined;
    subtotal?: number | undefined;
    tableNumber?: string | number | undefined;
    tax?: number | undefined;
    tip?: number | undefined;
    totalAmount?: number | undefined;
    table_number?: string | number | undefined;
    customer_name?: string | undefined;
    customer_phone?: string | undefined;
    customer_email?: string | undefined;
    seat_number?: number | undefined;
    total_amount?: number | undefined;
    seatNumber?: number | undefined;
    customerName?: string | undefined;
    customerEmail?: string | undefined;
    customerPhone?: string | undefined;
}>;